control de usuarios implementado

+++ src/services/api.services.js
@@ -1,8 +1,13 @@
+//'Bearer Zr2esjVc8O58XLNKLeHDaMNZGk_vFUwV'
 import { API } from "./conection";
+const getAuthorization =()=>{
+    const access_token = localStorage.length>0?localStorage.getItem('acces_token'):'';
+    return {authorization:`Bearer ${access_token}`}
+}
 const ApiServices = {
     getProductos:async (pagina)=>{
         let response ={};
-        await API.get('producto/view-pagination',{params:{pageSize:15,page:pagina},headers:{authorization:'Bearer Zr2esjVc8O58XLNKLeHDaMNZGk_vFUwV'}})
+        await API.get('producto/view-pagination',{params:{pageSize:15,page:pagina},headers:getAuthorization()})
             .then((res)=>{response=res})
             .catch((error)=>{
                 response = error.response?error.response:{};
@@ -14,7 +19,7 @@ const ApiServices = {
     
     registrarProducto:async (producto)=>{
         let response ={};
-        await API.post('producto/register',producto)
+        await API.post('producto/register',producto,{headers:getAuthorization()})
             .then((res)=>{response=res})
             .catch((error)=>{
                 response=error.response?error.response:{};
@@ -26,7 +31,7 @@ const ApiServices = {
 
     modificarProducto:async (id,cuerpo)=>{
         let response = {};
-        await API.put('producto/change-data?id='+id,cuerpo)
+        await API.put('producto/change-data?id='+id,cuerpo,{headers:getAuthorization()})
         .then((res)=>{response=res})
         .catch((error)=>{
             response=error.response?error.response:{};
@@ -36,7 +41,7 @@ const ApiServices = {
     },
     eliminarProducto:async(id)=>{
         let response = {};
-        await API.delete('producto/remove?id='+id)
+        await API.delete('producto/remove?id='+id,{headers:getAuthorization()})
         .then((res)=>{response=res})
         .catch((error)=>{
             response=error.response?error.response:{};
@@ -46,7 +51,7 @@ const ApiServices = {
     },
     getMarcas:async()=>{
         let response = {};
-        await API.get('marca/view-all-brands')
+        await API.get('marca/view-all-brands',{headers:getAuthorization()})
         .then((res)=>{response=res})
         .catch((error)=>{
             response = error.response?error.response:{};
@@ -56,7 +61,7 @@ const ApiServices = {
     },
     getSecciones:async()=>{
         let response = {};
-        await API.get('seccion/view-all-seccions')
+        await API.get('seccion/view-all-seccions',{headers:getAuthorization()})
         .then((res)=>{response=res})
         .catch((error)=>{
             response = error.response?error.response:{};
